+++ src/components/spinner/spinner.css
@@ -1,0 +1,10 @@
+.spinner {
+  position: absolute;
+  margin: auto;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  width: 100px;
+  height: 100px;
+}
