+++ src/reducers/spotify_reducer.js
@@ -15,7 +15,7 @@ const initialState = {
   error: null
 };
 
-export default function(state = initialState, action) {
+export default function (state = initialState, action) {
   switch (action.type) {
     case GET_USER_SPOTIFY_PROFILE:
       return { ...state, profile: action.payload };
