+++ src/reducers/tumblr_reducer.js
@@ -12,7 +12,7 @@ const initialState = {
   error: null
 };
 
-export default function(state = initialState, action) {
+export default function (state = initialState, action) {
   switch (action.type) {
     case GET_USER_TUMBLR_POSTS:
       return { ...state, tracks: action.payload };
